requirements-dev.txt
annotated-types==0.6.0
blis==0.7.11
camelot-py==0.9.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
cymem==2.0.8
et-xmlfile==1.1.0
fonttools==4.44.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.8.1
murmurhash==1.0.10
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
pdfminer.six==20221105
Pillow==10.1.0
preshed==3.0.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pyparsing==3.1.1
pypdf==3.17.0
PyPDF2==2.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tabulate==0.9.0
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.3
